survivor/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-environ==0.4.5
django-rest-swagger==2.2.0
django-support==0.2
djangorestframework==3.12.4
Events==0.4
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
limits==1.5.1
Markdown==3.3.4
MarkupSafe==2.0.1
openapi-codec==1.3.2
psycopg2-binary==2.9
pylogbeat==2.0.0
python-logstash==0.4.6
python-logstash-async==2.3.0
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5